refactor lanes query
