+++ scopes/component/component/ui/use-component.tsx
@@ -29,8 +29,8 @@ export function useComponent(host: string, id?: ComponentID): Component {
   const currentLane = lanesContext?.viewedLane;
   // when on a lane, always fetch all the logs starting from the 'head' version
   const laneComponentId = lanesContext?.viewedLane?.components.find(
-    (component) => component.model.id.fullName === targetId
-  )?.model.id;
+    (component) => component.id.fullName === targetId
+  )?.id;
 
   const componentIdStr = laneComponentId ? laneComponentId?.toString() : withVersion(targetId, version);
