+++ scopes/lanes/ui/hooks/lanes-query/index.ts
@@ -1,1 +1,3 @@
-export { useLanesQuery, useLaneComponentsQuery, useLaneReadmeQuery } from './lanes-query';
+export { useLaneComponents } from './use-lane-components';
+export { useLaneReadme } from './use-lane-readme';
+export { useLanes } from './use-lanes';
