+++ scopes/lanes/ui/hooks/lanes-query/use-lane-components/index.ts
@@ -1,0 +1,1 @@
+export { useLaneComponents } from './use-lane-components';
