+++ scopes/lanes/ui/hooks/lanes-query/use-lane-readme/index.ts
@@ -1,0 +1,1 @@
+export { useLaneReadme } from './use-lane-readme';
