+++ scopes/lanes/ui/hooks/lanes-query/use-lanes/index.ts
@@ -1,0 +1,1 @@
+export { useLanes } from './use-lanes';
